requirements.txt
requests==2.21.0
facebook-sdk==3.1.0